nfsd: fix cast to integer of different size
    ../../nfsd/loop.c: In function 'server_loop':
    ../../nfsd/loop.c:41:12: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

    ../../nfsd/main.c: In function 'create_server_thread':
    ../../nfsd/main.c:47:54: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
Message-ID: <20251005230901.17598-10-dnietoc@gmail.com>

+++ nfsd/loop.c
@@ -38,7 +38,7 @@
 void *
 server_loop (void *arg)
 {
-  int fd = (int) arg;
+  int fd = (intptr_t) arg;
   char buf[MAXIOSIZE];
   int xid;
   int *p, *r;
